feat: добавлена поддержка множественных callback в едином handler

+++ bot/middlewares/outer/statecontrol.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from typing import Any, Awaitable, Callable, Optional
+import logging
+from typing import Any, Awaitable, Callable
 
 from aiogram import BaseMiddleware
 from aiogram.fsm.context import FSMContext
@@ -8,23 +9,31 @@ from aiogram.types import CallbackQuery, Message, TelegramObject
 
 from bot.filters import ChatStates
 
+logger = logging.getLogger(__name__)
 
-class OuterMiddleware(BaseMiddleware):
+
+class StateControlMiddleware(BaseMiddleware):
     async def __call__(
         self,
         handler: Callable[[TelegramObject, dict[str, Any]], Awaitable[Any]],
         event: TelegramObject,
         data: dict[str, Any],
-    ) -> Optional[Any]:
+    ) -> Any:
         state: FSMContext = data["state"]
+
+        logger.debug("State: %s", await state.get_state())
 
         if isinstance(event, CallbackQuery):
             await state.set_state(ChatStates.ReadyToRespond)
         elif hasattr(event, "message") and isinstance(
-            getattr(event, "message", None), Message
+            getattr(event, "message"), Message
         ):
             message: Message = getattr(event, "message")
-            if message.text and message.text.startswith("/"):
+            if (
+                message.text
+                and message.text.startswith("/")
+                and message.text != "/start"
+            ):
                 await message.delete()
                 await state.set_state(ChatStates.ReadyToRespond)
